Junit/JunitProjects/SolutionProjects/AlljUnitLabs/CourseExamples/src/test/java/com/junit/exceptions/MentorTest.java
package com.junit.exceptions;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Test;

import com.junit.exceptions.Mentor;

class MentorTest {
	@Test
	void testFullName() {
		String expected = "Jane Doe";
		Mentor mentor = new Mentor("Jane", "Doe", 1);
		String actual = mentor.getFullName();
		assertEquals(expected, actual, "Full name should be Jane Doe");
	}
	
	@Test
	void testInvalidId() {
		assertThrows(IllegalArgumentException.class, () -> {
			new Mentor("Aadya", "Kumar", -1);
		});
	}
	
	@Test
	void testInvalidIdMessage() {
		Exception e = assertThrows(IllegalArgumentException.class, () -> {
			new Mentor("Aadya", "Kumar", -1);
		});
		assertEquals("id may not be negative", e.getMessage(), 
				"Expected appropriate message");
	}

}
